Add `Row::order`
